Start work on push token registration
